The big grid system switch patch WARNING: This renders GWC broken until further notice. Classes subject to change.
git-svn-id: http://geowebcache.org/svn/trunk@708 e7b91dd5-889a-44ae-8e97-0abfc27e49b2

+++ geowebcache/src/test/java/org/geowebcache/service/gmaps/GMapsConverterTest.java
@@ -20,8 +20,8 @@ public class GMapsConverterTest extends TestCase {
     public void testGMapsConverter() throws Exception {
     	/* Check origin location */
     	int x = 0; int y = 0; int z = 0;
-    	int[] gridLoc = GMapsConverter.convert(z, x, y);
-    	int[] solution = {0,0,0};
+    	long[] gridLoc = GMapsConverter.convert(z, x, y);
+    	long[] solution = {0,0,0};
     	assert(Arrays.equals(gridLoc, solution));
     	
     	/* Check zoomlevel */
